Change final project

+++ database/commentCrud.go
@@ -46,16 +46,20 @@ func GetSingleComment(commentID uint) (models.Comment, error) {
 	return comment, err
 }
 
-func UpdateComment(commentID uint, messageDto *dto.CommentMessage) (UpdatedAt time.Time, err error) {
-	comment, err := GetSingleComment(commentID)
+func UpdateComment(commentID, userID uint, messageDto *dto.CommentMessage) (comment models.Comment, err error) {
+	comment, err = GetSingleComment(commentID)
 	if err != nil {
+		return
+	}
+	if comment.UserID != userID {
+		comment = models.Comment{}
+		err = ErrIllegalUpdate
 		return
 	}
 	comment.Message = messageDto.Message
 	comment.UpdatedAt = time.Now()
 	err = db.Save(&comment).Error
 	if err == nil {
-		UpdatedAt = comment.UpdatedAt
 		log.Printf("Comment Updated: %+v\n", comment)
 	} else {
 		log.Printf("Failed Update Comment %+v with %+v because of %q\n", comment, messageDto, err.Error())
@@ -63,10 +67,13 @@ func UpdateComment(commentID uint, messageDto *dto.CommentMessage) (UpdatedAt ti
 	return
 }
 
-func DeleteCommentById(commentID uint) error {
+func DeleteComment(commentID, userID uint) error {
 	comment, err := GetSingleComment(commentID)
 	if err != nil {
 		return err
+	}
+	if comment.UserID != userID {
+		return ErrIllegalUpdate
 	}
 	if err := db.Delete(&comment, commentID).Error; err != nil {
 		return err
